+++ src/pyctpclient/__init__.py
@@ -1,3 +1,3 @@
 from ._ctpclient import *
-__version__ = "0.1.3a1"
+__version__ = "0.1.3a2"
 __author__ = "Holmes Conan"